@@ -63,4 +63,4 @@ Login.defaultProps = {
 };
 
 
-module.exports = connect(mapStateToProps, mapDispatchToProps)(Login);
+export default connect(mapStateToProps, mapDispatchToProps)(Login);
