@@ -11,5 +11,5 @@ export const youtubeParams = {
   },
 };
 
-// export const API_INDEX = 'http://localhost:4000';
-export const API_INDEX = 'https://api.balistos.com';
+export const API_INDEX = 'http://localhost:4000';
+// export const API_INDEX = 'https://api.balistos.com';
